requirements.txt
instaloader
twscrape
facebook-scraper
openai
asyncio
argparse
python-dotenv
pydantic
requests